chore:rename module, imports
