+++ src-tauri/core/mod.rs
@@ -1,4 +1,4 @@
 #[macro_use]
 pub mod macros;
-pub mod endpoints;
+pub mod handlers;
 pub mod transfer;
