+++ src-tauri/main.rs
@@ -1,12 +1,9 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use huehuehue::bindings;
-use huehuehue::core::endpoints::*;
-use huehuehue::huehuehue_handlers;
-use huehuehue::HueHueHue;
-use huehuehue::HueHueHueError;
-use huehuehue::HueHueHueState;
+use huehuehue::{
+    bindings, core::handlers::*, huehuehue_handlers, HueHueHue, HueHueHueError, HueHueHueState,
+};
 use log::info;
 use tauri::RunEvent;
 use tokio::sync::Mutex;
